fix isFull not working in springcloud and http application (#3212)
